+++ src/functions/authentication/structs.rs
