+++ src/functions/mod.rs
@@ -1,8 +1,8 @@
-pub mod fetch_stores;
-pub mod file_server;
+pub mod stores;
+pub mod files;
 pub mod logs;
 pub mod permissions;
-pub mod product_data;
-pub mod signing;
-pub mod user_control;
-pub mod version_check;
+pub mod products;
+pub mod authentication;
+pub mod users;
+pub mod versions;
