+++ src/functions/products/structs.rs
