+++ src/functions/stores/structs.rs
