+++ src/functions/users/structs.rs
