+++ src/routes/authentication.rs
@@ -3,8 +3,8 @@ use rocket::http::Status;
 use rocket::serde::json::Json;
 use rocket::{post, State};
 
-use crate::functions::signing::signin;
-use crate::functions::signing::structs::LoginParams;
+use crate::functions::authentication::signin;
+use crate::functions::authentication::structs::LoginParams;
 
 use crate::utils::structs::APIErrors;
