+++ src/server/request_guard/api_key.rs
@@ -1,6 +1,6 @@
 use rocket::{http::Status, request::{FromRequest, Outcome}, Request};
 
-use crate::functions::signing::validate_token;
+use crate::functions::authentication::validate_token;
 
 
 // Start Request Guard Functions
